% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conv.R
\name{temp_conv}
\alias{temp_conv}
\title{Temperature conversion from one scale to another}
\usage{
temp_conv(temp, start_scale, end_scale)
}
\arguments{
\item{temp}{Temperature value}

\item{start_scale}{Starting temperature scale}

\item{end_scale}{Ending temperature scale}
}
\value{
Temperature value in the end scale
}
\description{
Temperature conversion from one scale to another
}
\examples{
temp_conv(41, "F", "C")

}
